requirements.txt
openai>=1.35.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0